[base] [coding] Moved BWT and MTF from base/ to coding/.
