+++ coding/bwt_coder.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
+#include "coding/bwt.hpp"
 #include "coding/huffman.hpp"
+#include "coding/move_to_front.hpp"
 #include "coding/varint.hpp"
-
-#include "base/bwt.hpp"
-#include "base/move_to_front.hpp"
 
 #include <algorithm>
 #include <cstdint>
@@ -27,9 +26,9 @@ public:
                                   std::vector<uint8_t> & bwtBuffer)
   {
     bwtBuffer.resize(n);
-    auto const start = base::BWT(n, s, bwtBuffer.data());
+    auto const start = BWT(n, s, bwtBuffer.data());
 
-    base::MoveToFront mtf;
+    MoveToFront mtf;
     for (auto & b : bwtBuffer)
       b = mtf.Transform(b);
 
@@ -84,7 +83,7 @@ public:
     huffman.ReadAndDecode(source, std::back_inserter(bwtBuffer));
 
     size_t const n = bwtBuffer.size();
-    base::MoveToFront mtf;
+    MoveToFront mtf;
     for (size_t i = 0; i < n; ++i)
     {
       auto const b = mtf[bwtBuffer[i]];
@@ -96,7 +95,7 @@ public:
       CHECK_LESS(start, n, ());
 
     revBuffer.resize(n);
-    base::RevBWT(n, static_cast<size_t>(start), bwtBuffer.data(), revBuffer.data());
+    RevBWT(n, static_cast<size_t>(start), bwtBuffer.data(), revBuffer.data());
     return std::copy(revBuffer.begin(), revBuffer.end(), it);
   }
