+++ coding/move_to_front.hpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include <limits>
 
-namespace base
+namespace coding
 {
 class MoveToFront
 {
@@ -13,8 +13,8 @@ public:
 
   MoveToFront();
 
-  // Returns index of the byte |b| in the current sequence of bytes, then, moves |b|
-  // to the first positions.
+  // Returns index of the byte |b| in the current sequence of bytes,
+  // then moves |b| to the first position.
   uint8_t Transform(uint8_t b);
 
   uint8_t operator[](uint8_t i) const { return m_order[i]; }
@@ -22,4 +22,4 @@ public:
 private:
   std::array<uint8_t, kNumBytes> m_order;
 };
-}  // namespace
+}  // namespace coding
